Fix accel engine, implement repetition penalty + temp + etc

+++ indextts/infer_v2_5.py
@@ -153,7 +153,12 @@ class IndexTTS2:
                 use_deepspeed = False
                 print(f">> Failed to load DeepSpeed. Falling back to normal inference. Error: {e}")
 
-        self.gpt.post_init_gpt2_config(use_deepspeed=use_deepspeed, kv_cache=True, half=self.use_bf16)
+        self.gpt.post_init_gpt2_config(
+            use_deepspeed=use_deepspeed,
+            kv_cache=True,
+            half=self.use_bf16,
+            accel_dtype=self.dtype,
+        )
 
         if self.use_cuda_kernel:
             # preload the CUDA kernel for BigVGAN
